restructuring control and dispatch folders
