+++ h2integrate/control/control_rules/pyomo_rule_baseclass.py
@@ -1,7 +1,7 @@
 import openmdao.api as om
 
 
-class PyomoDispatchBaseClass(om.ExplicitComponent):
+class PyomoRuleBaseClass(om.ExplicitComponent):
     def initialize(self):
         self.options.declare("driver_config", types=dict)
         self.options.declare("plant_config", types=dict)
