+++ h2integrate/control/control_rules/storage/h2_storage.py
@@ -1,9 +1,9 @@
 import pyomo.environ as pyo
 
-from h2integrate.dispatch.dispatch_baseclass import PyomoDispatchBaseClass
+from h2integrate.control.control_rules.pyomo_rule_baseclass import PyomoRuleBaseClass
 
 
-class PyomoDispatchH2Storage(PyomoDispatchBaseClass):
+class PyomoDispatchH2Storage(PyomoRuleBaseClass):
     def compute(self, inputs, outputs, discrete_inputs, discrete_outputs):
         discrete_outputs["_create_variables"] = self._create_variables
         discrete_outputs["_create_port"] = self._create_port
